Show welcome page to only contributors
This commit moves the private routes into their own component beneath
and AuthGuard so they can access the user in the store.
